+++ app/src/main/java/com/masterwok/shrimplesearch/common/data/repositories/contracts/ConfigurationRepository.kt
@@ -2,5 +2,5 @@ package com.masterwok.shrimplesearch.common.data.repositories.contracts
 
 interface ConfigurationRepository {
     suspend fun incrementResultTapCount()
-    suspend fun getResultItemTapCount(): Int
+    suspend fun getResultItemTapCount(): Long
 }
